+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 @FetchPlan, @Limit and @Timeout extensions now modifies query instead of using OCommandRequest setters
 * (behavior change) @EdgeType and @VertexType extensions now assign superclass directly to annotated type (or not if hierarchy already contains required type)
 * Call activateOnCurrentThread() in pools for each connection obtain (e.g. context.getConnection()) to guarantee proper db instance bound to thread
+* Fix guice circular proxy (between transaction manager and pools)
 
 ### 3.1.1 (2015-08-16)
 * Improve graph connection recognition on repository methods: recognize all graph types implementing Vertex or Edge (e.g. OrientVertex, OrientEdge)